ENH: New qtTaskNode Design
Designed New qtTaskNode
-----------------------

This implementation only depends qtGraphicsItems and does not use Qt Widgets.
The result is both a cleaner looking design and also does not suffer from the
transformation problems that was noticed on Macs.

The new qtTaskNode is composed of 3 main graphics items:

* An item to display the Task's state as a color
* An item to display and manage the Task's name as well as the Task's activation status
* An item to control the Task's completed state

Supporting the Snapping of Port Nodes
-------------------------------------

It is now possible to snap a Task's Ports to be close to the Task itself.
This is controlled by the setSnapPorts and snapPorts methods on qtBaseTaskNode.

Other Changes
-------------
* Consolidated all of the color information used for task nodes and task port nodes into
  the qtDiagramViewConfiguration class
* When in connection mode, all single and double clink events are consumed by the diagram
* Fixed issue with Ports and Arcs not drawing correctly when initially reloaded into ModelBuilder
* Existing worklet and agent tests to use the old default style since they want to record task name changes