Fix FillOutAttributesAgent to handle definitions.
If attribute Definitions (but no explicit instances) are to be
validated, we must allow an `AttributeSet`'s `ResourceAttributes`
to be empty without forcing the agent state to be unavailable.
To do this, we must note when a new attribute resource is added
and update the internal state of the agent even if no attribute
instances are added to the `AttributeSet`.