Track the project manager in the attribute panel.
In order to properly discern when an attribute resource should be
marked as "displayable" by the panel (i.e., when updating the
`smtk.attribute_panel.display_hint` property), we need to check
whether the attribute resource is owned by a project. This in turn
means we need access to the project manager which was not previously
added to the qtUIManager instance used to configure the view.