CMake fixes for Windows.
Adds `/bigobj` to resolve build issue on Windows when compiling in Debug.
Improved logic for copy boost::filesystem DLL to consider the current build type so that the correct version is copied to the bin folder. This resolves an issue that arises when both the release version and the debug version are present.