Provide handler functions for operation instances.
You can now have a handler function called once an operation
completes (as if it were an observer but only on the instance
of the operation you add it to, and only for one invocation).

This also removes the wrapping of many copy-constructors and
assignment operations in python because (a) they never should
have been added and (b) the mutex guarding the container of
handlers does not provide a default copy constructor which
now prohibits them from existing.